refactor(api): make routes thin and use entities
