+++ src/app/api/brands/route.js
@@ -1,13 +1,7 @@
-import { getDB } from '@/lib/db';
-import { getAllBrands } from '@/lib/queries/brands';
-import { ErrorHandler } from '@/lib/utils/errorHandler';
+import { getAllBrands } from '@/entities/brand';
+import { ErrorHandler } from '@/shared/lib';
 
 export const GET = ErrorHandler(async () => {
-  const db = await getDB();
-  console.log('[API] Connection to the database was successful');
-
-  const [rows] = await db.query(getAllBrands);
-  console.log('[API] query result:', rows);
-
+  const rows = await getAllBrands();
   return Response.json(rows);
 });
