+++ src/app/api/reviews/route.js
@@ -1,13 +1,5 @@
-import { getDB } from '@/lib/db';
-import {
-  getReviewsAvgByProduct,
-  getReviewsCount,
-  getReviewsCountByProduct,
-  getReviewsPage,
-  getReviewsPageByProduct,
-  insertReview,
-} from '@/lib/queries/reviews';
-import { ErrorHandler } from '@/lib/utils/errorHandler';
+import { createReview, getReviewsByProduct, getReviewsFeed } from '@/entities/review';
+import { ErrorHandler } from '@/shared/lib';
 
 export const POST = ErrorHandler(async (req) => {
   const { productId, rating, comment, authorName } = await req.json();
@@ -22,8 +14,7 @@ export const POST = ErrorHandler(async (req) => {
   const name = (authorName || '').trim().slice(0, 60) || null;
   const text = (comment || '').trim().slice(0, 1000) || null;
 
-  const db = await getDB();
-  await db.execute(insertReview, [pid, r, name, text]);
+  await createReview({ productId: pid, rating: r, authorName: name, comment: text });
 
   return new Response(null, { status: 201 });
 });
@@ -35,34 +26,17 @@ export const GET = ErrorHandler(async (req) => {
   const productId = productIdRaw ? Number(productIdRaw) : null;
 
   const page = Math.max(Number(searchParams.get('page')) || 1, 1);
-  const limitRaw = Number(searchParams.get('limit')) || 10;
-  const limit = Math.min(Math.max(limitRaw, 5), 50);
-  const offset = (page - 1) * limit;
-
-  const db = await getDB();
+  const limit = Number(searchParams.get('limit')) || 10;
 
   if (productId !== null) {
     if (!Number.isFinite(productId) || productId <= 0) {
       return new Response(JSON.stringify({ error: 'Invalid productId' }), { status: 400 });
     }
 
-    const [[countRow]] = await db.query(getReviewsCountByProduct, [productId]);
-    const total = Number(countRow?.total) || 0;
-    const totalPages = Math.max(Math.ceil(total / limit), 1);
-
-    const [[avgRow]] = await db.query(getReviewsAvgByProduct, [productId]);
-    const avgRating = Number(avgRow?.avgRating) || 0;
-
-    const [rows] = await db.query(getReviewsPageByProduct, [productId, limit, offset]);
-
-    return Response.json({ items: rows, total, totalPages, page, limit, avgRating });
+    const data = await getReviewsByProduct({ productId, page, limit });
+    return Response.json(data);
   }
 
-  const [[countRow]] = await db.query(getReviewsCount);
-  const total = Number(countRow?.total) || 0;
-  const totalPages = Math.max(Math.ceil(total / limit), 1);
-
-  const [rows] = await db.query(getReviewsPage, [offset, limit]);
-
-  return Response.json({ items: rows, total, totalPages, page, limit });
+  const data = await getReviewsFeed({ page, limit });
+  return Response.json(data);
 });
